+++ etc/stir-prepare.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+#
+# .stir-prepare.sh
+#
+MYSQL_USER="someuser"
+MYSQL_PASS="somepassword"